+++ src/test/scala/com/fulcrumgenomics/bam/IdentifyPrimersTest.scala
@@ -35,6 +35,7 @@ import htsjdk.samtools.util.SequenceUtil
 import org.scalatest.OptionValues
 import com.fulcrumgenomics.FgBioDef.unreachable
 import com.fulcrumgenomics.bam.IdentifyPrimers._
+import com.fulcrumgenomics.bam.identifyprimers.IdentifyPrimers
 import com.fulcrumgenomics.commons.io.PathUtil
 import com.fulcrumgenomics.util.Metric
 import htsjdk.samtools.SamPairUtil
